\name{rgw-package}
\alias{rgw-package}
\alias{rgw}
\docType{package}
\title{
Goodman-Weare Affine-Invariant Sampling
}
\description{
This package implements the affine-invariant method of Goodman & Weare (2010) <DOI:10.2140/camcos.2010.5.65>, a method of producing Monte-Carlo samples from a target distribution. The implementation is based on the description of the `emcee' python package (implementing the same method) written by Forman-Mackey et al. (2012) <DOI:10.1086/670067>. See `References' in the documentation of the GoodmanWeare function for full citation details.
}
\details{
\tabular{ll}{
Package: \tab rgw\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2016-10-10\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
}

}
\author{
Adam Mantz <amantz@slac.stanford.edu>
}

\keyword{ package }
